Refactor: Extract ExternalMetadataLoader to separate module
- Created app/utils/external_metadata_loaders.py to house the ExternalMetadataLoader class
- Moved metadata loading methods (_emit_db_error, _validate_db_path, _load_json_file, etc.) from MetadataManager to ExternalMetadataLoader
- Updated MetadataManager to instantiate and use ExternalMetadataLoader instance
- Removed duplicate constants and imports from metadata.py
- Improved code organization by separating external metadata loading concerns
